+++ app/api/clubs/route.ts
@@ -32,6 +32,7 @@ export async function GET() {
 			`
 				)
 				.eq("user_id", user?.id || "")
+				.order("is_favorite", { ascending: false })
 
 			if (error) {
 				throw error