+++ components/ui/books/club/spreads/dashboard/dashboard-page-left.tsx
@@ -33,6 +33,7 @@ export function DashboardPageLeft({ userSpreadIndex }: Props) {
 
 	return (
 		<MotionCard
+			id={`club-${clubMembership?.club.id}-dashboard-page-left`}
 			className="bg-background flex-1 h-1/2 md:h-full md:w-1/2 relative border-b-0 rounded-b-none md:border-b md:rounded-b-lg md:border-r-0 md:rounded-tr-none md:rounded-br-none shadow-shadow shadow-md"
 			variants={leftVariants}
 			initial="initial"